+++ internal/provider/data_source_iosxe_interface_ethernet.go
@@ -519,7 +519,7 @@ func (d *InterfaceEthernetDataSource) Schema(ctx context.Context, req datasource
 				MarkdownDescription: "Assign policy-map to the output of an interface",
 				Computed:            true,
 			},
-			"ip_flow_monitor": schema.ListNestedAttribute{
+			"ip_flow_monitors": schema.ListNestedAttribute{
 				MarkdownDescription: "Apply a Flow Monitor",
 				Computed:            true,
 				NestedObject: schema.NestedAttributeObject{